pyproject.toml
[tool.poetry]
name = "zeuscloud-iamspy"
version = "0.7.0"
description = "ZeusCloud fork of IAMSpy"
authors = ["Varun Jain <varun@zeuscloud.io>"]

[tool.poetry.dependencies]
python = "^3.8"
z3-solver = "*"
pydantic = "*"
typer = "*"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
ipython = "*"
black = "^22.3.0"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
pdbpp = "^0.10.3"

[tool.poetry.scripts]
iamspy = 'zeuscloud_iamspy.cli:app'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
